Strongly type ResourceMetadata and ResourceKind
