+++ bound/typescript/src/resources/offering.ts
@@ -1,8 +1,8 @@
+import { ResourceMetadata } from ".";
 import { BearerDid } from "../bearer-did";
 import { tbdexError } from "../errors";
 import { PresentationDefinition } from "../presentation-definition";
 import wasm from "../wasm";
-import { ResourceMetadata } from "../wasm/generated-mappings";
 
 export class Offering {
   readonly metadata: ResourceMetadata;
